# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"TBScreenAI API","description":"Shared backend for TBScreenAI (roles: doctor, admin_rs, super_admin). All medical data is tenant-scoped per hospital. AI inference is MOCKED in this phase.","version":"0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenPair"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh","operationId":"refresh_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenPair"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patients":{"get":{"tags":["patients"],"summary":"List Patients","operationId":"list_patients_api_v1_patients_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PatientOut"},"title":"Response List Patients Api V1 Patients Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["patients"],"summary":"Create Patient","operationId":"create_patient_api_v1_patients_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patients/{patient_id}":{"get":{"tags":["patients"],"summary":"Get Patient","operationId":"get_patient_api_v1_patients__patient_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Patient Id"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["patients"],"summary":"Update Patient","operationId":"update_patient_api_v1_patients__patient_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Patient Id"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["patients"],"summary":"Delete Patient","operationId":"delete_patient_api_v1_patients__patient_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Patient Id"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/diagnoses/infer":{"post":{"tags":["diagnoses"],"summary":"Infer","description":"Accepts a chest X-ray image and returns a structured MOCK result.\n\nThe image is validated but not stored yet (storage lands with the real\nmodel integration).","operationId":"infer_api_v1_diagnoses_infer_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_infer_api_v1_diagnoses_infer_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/diagnoses":{"get":{"tags":["diagnoses"],"summary":"List Diagnoses","operationId":"list_diagnoses_api_v1_diagnoses_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"patient_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Patient Id"}},{"name":"status_filter","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status Filter"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiagnosisOut"},"title":"Response List Diagnoses Api V1 Diagnoses Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["diagnoses"],"summary":"Create Diagnosis","operationId":"create_diagnosis_api_v1_diagnoses_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosisCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosisOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/diagnoses/{diagnosis_id}":{"get":{"tags":["diagnoses"],"summary":"Get Diagnosis","operationId":"get_diagnosis_api_v1_diagnoses__diagnosis_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"diagnosis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Diagnosis Id"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosisOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/diagnoses/{diagnosis_id}/status":{"patch":{"tags":["diagnoses"],"summary":"Update Status","description":"Doctor validation verdict — mirrors the Flutter ValidationScreen flow.","operationId":"update_status_api_v1_diagnoses__diagnosis_id__status_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"diagnosis_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Diagnosis Id"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosisStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagnosisOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sync/push":{"post":{"tags":["sync"],"summary":"Push","description":"Apply offline changes from the tablet. Idempotent per client_op_id.","operationId":"push_api_v1_sync_push_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncPushRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncPushResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sync/pull":{"get":{"tags":["sync"],"summary":"Pull","description":"Server changes for this hospital since `since` (full snapshot if omitted),\nplus the latest AI model version.","operationId":"pull_api_v1_sync_pull_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"since","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Since"}},{"name":"x-tenant-id","in":"header","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncPullResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sync/model-version":{"get":{"tags":["sync"],"summary":"Latest Model Version","description":"Latest AI model release — powers the ModelUpdateCard check.","operationId":"latest_model_version_api_v1_sync_model_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/ModelVersionOut"},{"type":"null"}],"title":"Response Latest Model Version Api V1 Sync Model Version Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"Body_infer_api_v1_diagnoses_infer_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_infer_api_v1_diagnoses_infer_post"},"DiagnosisCreate":{"properties":{"patient_id":{"type":"string","format":"uuid","title":"Patient Id"},"is_positive":{"type":"boolean","title":"Is Positive"},"confidence":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Confidence"},"model_version":{"type":"string","maxLength":50,"title":"Model Version"},"processing_time_ms":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Processing Time Ms"},"findings":{"$ref":"#/components/schemas/Findings"},"image_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Path"},"diagnosed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Diagnosed At"}},"type":"object","required":["patient_id","is_positive","confidence","model_version","findings"],"title":"DiagnosisCreate"},"DiagnosisOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"tenant_id":{"type":"string","format":"uuid","title":"Tenant Id"},"patient_id":{"type":"string","format":"uuid","title":"Patient Id"},"created_by":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Created By"},"is_positive":{"type":"boolean","title":"Is Positive"},"confidence":{"type":"integer","title":"Confidence"},"model_version":{"type":"string","title":"Model Version"},"processing_time_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Processing Time Ms"},"findings":{"$ref":"#/components/schemas/Findings"},"status":{"type":"string","title":"Status"},"doctor_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doctor Note"},"image_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Path"},"diagnosed_at":{"type":"string","format":"date-time","title":"Diagnosed At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","tenant_id","patient_id","created_by","is_positive","confidence","model_version","processing_time_ms","findings","status","doctor_note","image_path","diagnosed_at","created_at","updated_at"],"title":"DiagnosisOut"},"DiagnosisStatusUpdate":{"properties":{"status":{"type":"string","pattern":"^(pending|agreed|disagreed)$","title":"Status"},"doctor_note":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Doctor Note"}},"type":"object","required":["status"],"title":"DiagnosisStatusUpdate"},"Findings":{"properties":{"consolidation":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Consolidation"},"cavity":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Cavity"},"effusion":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Effusion"},"fibrotic":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Fibrotic"},"calcification":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Calcification"}},"type":"object","required":["consolidation","cavity","effusion","fibrotic","calcification"],"title":"Findings","description":"Finding percentages — mirrors the Flutter ValidationFindings contract."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InferenceResult":{"properties":{"is_positive":{"type":"boolean","title":"Is Positive"},"confidence":{"type":"integer","title":"Confidence"},"processing_time_ms":{"type":"integer","title":"Processing Time Ms"},"model_version":{"type":"string","title":"Model Version"},"findings":{"$ref":"#/components/schemas/Findings"},"is_mock":{"type":"boolean","title":"Is Mock","default":true}},"type":"object","required":["is_positive","confidence","processing_time_ms","model_version","findings"],"title":"InferenceResult","description":"Structured MOCK inference response.\n\nReal model integration is out of scope — the contract is what matters:\nit mirrors the Flutter DiagnosisOutcome so the tablet can render it as-is."},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"ModelVersionOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"version":{"type":"string","title":"Version"},"file_size_mb":{"type":"number","title":"File Size Mb"},"release_date":{"type":"string","format":"date","title":"Release Date"},"changelog":{"items":{"type":"string"},"type":"array","title":"Changelog"},"is_latest":{"type":"boolean","title":"Is Latest"},"download_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Download Url"}},"type":"object","required":["id","version","file_size_mb","release_date","changelog","is_latest","download_url"],"title":"ModelVersionOut"},"PatientCreate":{"properties":{"code":{"type":"string","maxLength":20,"title":"Code"},"name":{"type":"string","maxLength":255,"title":"Name"},"age":{"type":"integer","maximum":150.0,"minimum":0.0,"title":"Age"},"gender":{"type":"string","pattern":"^(Male|Female)$","title":"Gender"},"status":{"type":"string","pattern":"^(Normal|Positive|Suspected)$","title":"Status","default":"Normal"},"confidence":{"anyOf":[{"type":"integer","maximum":100.0,"minimum":0.0},{"type":"null"}],"title":"Confidence"},"last_visit":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Visit"},"history":{"items":{"type":"string"},"type":"array","title":"History"}},"type":"object","required":["code","name","age","gender"],"title":"PatientCreate"},"PatientOut":{"properties":{"code":{"type":"string","maxLength":20,"title":"Code"},"name":{"type":"string","maxLength":255,"title":"Name"},"age":{"type":"integer","maximum":150.0,"minimum":0.0,"title":"Age"},"gender":{"type":"string","pattern":"^(Male|Female)$","title":"Gender"},"status":{"type":"string","pattern":"^(Normal|Positive|Suspected)$","title":"Status","default":"Normal"},"confidence":{"anyOf":[{"type":"integer","maximum":100.0,"minimum":0.0},{"type":"null"}],"title":"Confidence"},"last_visit":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Visit"},"history":{"items":{"type":"string"},"type":"array","title":"History"},"id":{"type":"string","format":"uuid","title":"Id"},"tenant_id":{"type":"string","format":"uuid","title":"Tenant Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["code","name","age","gender","id","tenant_id","created_at","updated_at"],"title":"PatientOut"},"PatientUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Name"},"age":{"anyOf":[{"type":"integer","maximum":150.0,"minimum":0.0},{"type":"null"}],"title":"Age"},"gender":{"anyOf":[{"type":"string","pattern":"^(Male|Female)$"},{"type":"null"}],"title":"Gender"},"status":{"anyOf":[{"type":"string","pattern":"^(Normal|Positive|Suspected)$"},{"type":"null"}],"title":"Status"},"confidence":{"anyOf":[{"type":"integer","maximum":100.0,"minimum":0.0},{"type":"null"}],"title":"Confidence"},"last_visit":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Last Visit"},"history":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"History"}},"type":"object","title":"PatientUpdate"},"RefreshRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshRequest"},"SyncPullResponse":{"properties":{"server_time":{"type":"string","format":"date-time","title":"Server Time"},"patients":{"items":{"$ref":"#/components/schemas/PatientOut"},"type":"array","title":"Patients"},"diagnoses":{"items":{"$ref":"#/components/schemas/DiagnosisOut"},"type":"array","title":"Diagnoses"},"latest_model":{"anyOf":[{"$ref":"#/components/schemas/ModelVersionOut"},{"type":"null"}]}},"type":"object","required":["server_time","patients","diagnoses","latest_model"],"title":"SyncPullResponse"},"SyncPushItem":{"properties":{"client_op_id":{"type":"string","format":"uuid","title":"Client Op Id"},"entity_type":{"type":"string","enum":["patient","diagnosis"],"title":"Entity Type"},"operation":{"type":"string","enum":["create","update"],"title":"Operation"},"entity_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Entity Id"},"base_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Base Updated At"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"}},"type":"object","required":["client_op_id","entity_type","operation","payload"],"title":"SyncPushItem"},"SyncPushRequest":{"properties":{"device_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Device Id"},"items":{"items":{"$ref":"#/components/schemas/SyncPushItem"},"type":"array","maxItems":500,"title":"Items"}},"type":"object","required":["items"],"title":"SyncPushRequest"},"SyncPushResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/SyncPushResult"},"type":"array","title":"Results"}},"type":"object","required":["results"],"title":"SyncPushResponse"},"SyncPushResult":{"properties":{"client_op_id":{"type":"string","format":"uuid","title":"Client Op Id"},"status":{"type":"string","enum":["applied","skipped","conflict"],"title":"Status"},"entity_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Entity Id"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail"}},"type":"object","required":["client_op_id","status"],"title":"SyncPushResult"},"TokenPair":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserOut"}},"type":"object","required":["access_token","refresh_token","user"],"title":"TokenPair"},"UserOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"type":"string","title":"Role"},"tenant_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Tenant Id"}},"type":"object","required":["id","email","full_name","role","tenant_id"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}